+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Support for custom networks
 - When generating new mnemonic, show it without a trace.
 - Support import slashing protection interchange data [EIP-3076](https://eips.ethereum.org/EIPS/eip-3076)
   while running the setup with the `cli` command or with the new command `slashing-import`.